+++ console/input.go
@@ -1,4 +1,4 @@
-package presentation
+package console
 
 import (
 	"errors"
